web/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.11.7
certifi==2021.5.30
charset-normalizer==2.0.12
datacache==1.2.1
dill==0.3.4
Django==3.2.18
django-crispy-forms==1.13.0
django-rq==2.7.0
django-extensions==3.2.1
gunicorn==20.1.0
gtfparse==1.3.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
memoized-property==1.0.3
mock==5.0.1
nose==1.3.7
numpy==1.19.5
openpyxl==3.1.1
pandas==1.4
platformdirs==2.4.0
plotly==5.13.0
primer3-py==0.6.1
progressbar33==2.4
psycopg2==2.9.5
pyensembl==2.2.4
pylint==2.13.9
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.27.1
serializable==0.2.1
simplejson==3.18.3
six==1.16.0
SQLAlchemy==1.4.47
sqlparse==0.4.3
tenacity==8.2.1
tinytimer==0.0.0
tomli==1.2.3
typechecks==0.1.0
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.14
wrapt==1.14.1
biopython==1.79